Log users out on 401 errors (#174)
If for whatever reason we are unable to authenticate the user, in many
cases thanks to our long-lived credentials hack, users can become stuck
and unable to recover without manually removing their credentials. This
PR detects 401 errors and logs the user out.

It'd be nice to have uniform error handling in place rather than
displaying a white page of death when unexpected things happen, but this
is better than what we have now. I considered trying to wrap such errors
in a custom type and catch them somewhere else. The complexity didn't
seem warranted, for now.