Back off connection attempts when failure is post-open (#473)
If something goes wrong in the sync-response handshake, the backend may
send a close signal which prompts the client to reconnect. This occurs
after onopen is called. As a result, we enter into a rapid fire
reconnect sequence, connecting every 500 milliseconds with no backoff.

This PR addresses that by maintaining the backoff time even when a
socket is opened successfully.