Avoid throwing on doc updates when operating in offline mode (#485)
When loading the page in offline mode, no socket will have opened
successfully and thus no clientMessageHandler will be set. Doc
modifications trigger the on-update observer which calls
#clientMessageHandler and in turn throws.

There isn't much we can do other than emit a debug log and move on. The
update will be persisted to indexed DB and sync'd on next connect.