Introduce and wire up Rust yproxy for interacting with yrs Doc (#501)
Changes:
- Introduce Rust yproxy objects for interacting with yrs which includes
wrappers for the Graph MapRef and the Task MapRef that provide simple
type safe getters and setters.
- Setters try to do the optimal mutation. For example, don't do anything
at all if the value hasn't changed. This is prominent when mutating
children. We use the `similar` library to compute a sequence of
mutations that transforms the previous value into the new value.
- Migrate the import and export endpoints to use the new yproxy objects
and remove the old doc endpoint which is no longer needed – use the
export endpoint instead.