Create tasks from Github PRs (#594)
* Adds a new `plugin_configs` DB table
* Exposes a `plugins/github/app/webhook` endpoint that receives push
notifications from installations
* Exposes a `plugins/github/poll` endpoint in dev mode to allow
on-demand polling
* Runs polling in the background every 16 minutes

Task are created under a task with a fixed ID, `plugin_github`, with a
new `kind` field set to `github` and `url` field set to the PRs URL.

The poller looks for children of the given parent matching the given
kind, fetch PRs from github and then creates/updates/resolves tasks. The
webhook looks for OPENED, CLOSED and EDITED PR updates and does the
same.

In both cases, the target project is determined based on the stored
configuration associated with the Github app installation. We may later
want to adjust this to allow users to map specific repositories onto
more than one project rather than map everything into a single project.