Avoid waiting for slow clients to respond (#1025)
On shutdown we send a close frame to all clients and then wait, with a
timeout, for them to respond with their own close frame. It's not
uncommon for some clients to be latent and take quite awhile to respond.
This delays shutdown and causes unavailability. We can avoid this by NOT
waiting for the close frame response from clients.